fix: The way elements are created and new design patterns were added as factory and startegy
