+++ src/helpers/opcionesEntidad.js
@@ -1,0 +1,9 @@
+const opcionesEntidad = {
+    PK: {type:'PK', color: "#5B0000"},
+    FK: {type:'FK', color: "#5B0000"},
+    ATRIBUTO: {type:'ATRIBUTO', color: "#097461"},
+    RELACION: {type:'RELACION'},
+    TXT: {type:'TXT'}
+}
+
+export default opcionesEntidad;
